Revert "Moves the eloquent contract tests to a trait"
This reverts commit a7e57c84ef8036e3dfb052d073943b1e9ad6ceb0.
