+++ tests/Integration/Database/SqlServer/DatabaseEloquentSqlServerIntegrationTest.php
@@ -4,15 +4,11 @@ namespace Illuminate\Tests\Integration\Database\SqlServer;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
-use Illuminate\Tests\Integration\Database\WithEloquentIntegrationTests;
 
 class DatabaseEloquentSqlServerIntegrationTest extends SqlServerTestCase
 {
-    use WithEloquentIntegrationTests;
-
-    protected $eloquentModelClass = DatabaseEloquentSqlServerIntegrationUser::class;
-
     protected function defineDatabaseMigrationsAfterDatabaseRefreshed()
     {
         if (! Schema::hasTable('database_eloquent_sql_server_integration_users')) {
@@ -28,6 +24,55 @@ class DatabaseEloquentSqlServerIntegrationTest extends SqlServerTestCase
     protected function destroyDatabaseMigrations()
     {
         Schema::drop('database_eloquent_sql_server_integration_users');
+    }
+
+    public function testCreateOrFirst()
+    {
+        $user1 = DatabaseEloquentSqlServerIntegrationUser::createOrFirst(['email' => 'taylorotwell@gmail.com']);
+
+        $this->assertSame('taylorotwell@gmail.com', $user1->email);
+        $this->assertNull($user1->name);
+
+        $user2 = DatabaseEloquentSqlServerIntegrationUser::createOrFirst(
+            ['email' => 'taylorotwell@gmail.com'],
+            ['name' => 'Taylor Otwell']
+        );
+
+        $this->assertEquals($user1->id, $user2->id);
+        $this->assertSame('taylorotwell@gmail.com', $user2->email);
+        $this->assertNull($user2->name);
+
+        $user3 = DatabaseEloquentSqlServerIntegrationUser::createOrFirst(
+            ['email' => 'abigailotwell@gmail.com'],
+            ['name' => 'Abigail Otwell']
+        );
+
+        $this->assertNotEquals($user3->id, $user1->id);
+        $this->assertSame('abigailotwell@gmail.com', $user3->email);
+        $this->assertSame('Abigail Otwell', $user3->name);
+
+        $user4 = DatabaseEloquentSqlServerIntegrationUser::createOrFirst(
+            ['name' => 'Dries Vints'],
+            ['name' => 'Nuno Maduro', 'email' => 'nuno@laravel.com']
+        );
+
+        $this->assertSame('Nuno Maduro', $user4->name);
+    }
+
+    public function testCreateOrFirstWithinTransaction()
+    {
+        $user1 = DatabaseEloquentSqlServerIntegrationUser::createOrFirst(['email' => 'taylor@laravel.com']);
+
+        DB::transaction(function () use ($user1) {
+            $user2 = DatabaseEloquentSqlServerIntegrationUser::createOrFirst(
+                ['email' => 'taylor@laravel.com'],
+                ['name' => 'Taylor Otwell']
+            );
+
+            $this->assertEquals($user1->id, $user2->id);
+            $this->assertSame('taylor@laravel.com', $user2->email);
+            $this->assertNull($user2->name);
+        });
     }
 }
