--- tests/Integration/Database/WithEloquentIntegrationTests.php
@@ -1,57 +1,0 @@
-<?php
-
-namespace Illuminate\Tests\Integration\Database;
-
-use Illuminate\Support\Facades\DB;
-
-trait WithEloquentIntegrationTests
-{
-    public function testCreateOrFirst()
-    {
-        $user1 = $this->eloquentModelClass::createOrFirst(['email' => 'taylorotwell@gmail.com']);
-
-        $this->assertSame('taylorotwell@gmail.com', $user1->email);
-        $this->assertNull($user1->name);
-
-        $user2 = $this->eloquentModelClass::createOrFirst(
-            ['email' => 'taylorotwell@gmail.com'],
-            ['name' => 'Taylor Otwell']
-        );
-
-        $this->assertEquals($user1->id, $user2->id);
-        $this->assertSame('taylorotwell@gmail.com', $user2->email);
-        $this->assertNull($user2->name);
-
-        $user3 = $this->eloquentModelClass::createOrFirst(
-            ['email' => 'abigailotwell@gmail.com'],
-            ['name' => 'Abigail Otwell']
-        );
-
-        $this->assertNotEquals($user3->id, $user1->id);
-        $this->assertSame('abigailotwell@gmail.com', $user3->email);
-        $this->assertSame('Abigail Otwell', $user3->name);
-
-        $user4 = $this->eloquentModelClass::createOrFirst(
-            ['name' => 'Dries Vints'],
-            ['name' => 'Nuno Maduro', 'email' => 'nuno@laravel.com']
-        );
-
-        $this->assertSame('Nuno Maduro', $user4->name);
-    }
-
-    public function testCreateOrFirstWithinTransaction()
-    {
-        $user1 = $this->eloquentModelClass::createOrFirst(['email' => 'taylor@laravel.com']);
-
-        DB::transaction(function () use ($user1) {
-            $user2 = $this->eloquentModelClass::createOrFirst(
-                ['email' => 'taylor@laravel.com'],
-                ['name' => 'Taylor Otwell']
-            );
-
-            $this->assertEquals($user1->id, $user2->id);
-            $this->assertSame('taylor@laravel.com', $user2->email);
-            $this->assertNull($user2->name);
-        });
-    }
-}
